- Implements logout endpoint - Restructure front-end environment
> Note that restructuring has involved moving constants, particularly page routing to a config file. `src\handlers` directory has been added, which contain the functions associated with initiating specific events
