+++ src/App.js
@@ -1,41 +1,42 @@
 // App.js
 import React from 'react';
-import { BrowserRouter as Router, Route, Switch} from 'react-router-dom';
-import Landing from './Landing';
-import UserProfile from './UserProfile';
-import Experience from './Experience';
-import Application from './Application';
-import Registration from './Registration';
-import Login from './Login';
-import Error from './Error';
+import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
+import Landing from './pages/Landing';
+import UserProfile from './pages/UserProfile';
+import Experience from './pages/Experience';
+import JobApplication from './pages/JobApplication';
+import Registration from './pages/Registration';
+import Login from './pages/Login';
+import Error from './pages/Error';
+import { PATHS } from './config/pageConfig'; // Import PATHS
 
 function App() {
   return (
-    <Router basename = "/hire-wire-front-end">
-      <Switch>
-        <Route exact path="/">
-          <Landing />
-        </Route>
-        <Route exact path="/login">
-          <Login />
-        </Route>
-        <Route exact path="/userprofile">
-          <UserProfile />
-        </Route>
-        <Route exact path="/experience">
-          <Experience />
-        </Route>
-        <Route exact path="/application">
-          <Application />
-        </Route>
-        <Route exact path="/registration">
-          <Registration />
-        </Route>
-        <Route exact path="/error">
-          <Error />
-        </Route>
-      </Switch>
-    </Router>
+      <Router basename="/hire-wire-front-end">
+        <Switch>
+          <Route exact path={PATHS.HOME}>
+            <Landing />
+          </Route>
+          <Route exact path={PATHS.LOGIN}>
+            <Login />
+          </Route>
+          <Route exact path={PATHS.USER_PROFILE}>
+            <UserProfile />
+          </Route>
+          <Route exact path={PATHS.EXPERIENCE}>
+            <Experience />
+          </Route>
+          <Route exact path={PATHS.JOB_APPLICATION}>
+            <JobApplication />
+          </Route>
+          <Route exact path={PATHS.REGISTRATION}>
+            <Registration />
+          </Route>
+          <Route exact path={PATHS.ERROR}>
+            <Error />
+          </Route>
+        </Switch>
+      </Router>
   );
 }
