+++ src/config/pageConfig.js
@@ -1,0 +1,11 @@
+// config/pageConfig.js
+
+export const PATHS = {
+    HOME: '/',
+    LOGIN: '/login',
+    USER_PROFILE: '/userprofile',
+    EXPERIENCE: '/experience',
+    JOB_APPLICATION: '/jobapplication',
+    REGISTRATION: '/registration',
+    ERROR: '/error',
+};
