+++ src/handlers/userProfileHandlers.js
@@ -1,0 +1,13 @@
+// handlers/userProfileHandlers.js
+
+// Save user profile changes (placeholder for actual save logic)
+export const handleSaveUserProfileClick = () => {
+    // Implement save logic here
+    console.log("User profile saved!");
+};
+
+// Delete user profile (placeholder for actual delete logic)
+export const handleDeleteProfileClick = () => {
+    // Implement delete logic here
+    console.log("User profile deleted!");
+};
