+++ src/templates/Landing.css
@@ -7,9 +7,9 @@ body {
     display: flex;
     justify-content: center;
     align-items: center;
-    background: url('login_images/login_page.jpg') no-repeat center center fixed;
+    background: url('../login_images/login_page.jpg') no-repeat center center fixed;
   }
-  
+
   .landing-container {
     display: flex;
     flex-direction: column;
@@ -17,23 +17,23 @@ body {
     justify-content: center;
     width: 100%;
     height: 100%;
-    background: url('login_images/login_page.png') no-repeat center center fixed;
+    background: url('../login_images/login_page.png') no-repeat center center fixed;
     background-size: cover;
   }
-  
+
   .landing-container h1 {
     font-size: 4rem;
     color: #000000;
     margin-bottom: 2rem;
   }
-  
+
   .button-group {
     display: flex;
     flex-direction: column;
     align-items: center;
     gap: 1rem;
   }
-  
+
   .login-button, .register-button {
     width: 250px;
     padding: 1rem 0;
@@ -44,22 +44,21 @@ body {
     cursor: pointer;
     transition: background-color 0.3s ease;
   }
-  
+
   .login-button {
     background-color: #3B82F6; /* Blue color for Log In button */
     color: #ffffff;
   }
-  
+
   .login-button:hover {
     background-color: #2563EB;
   }
-  
+
   .register-button {
     background-color: #333333; /* Dark color for Register button */
     color: #ffffff;
   }
-  
+
   .register-button:hover {
     background-color: #444444;
   }
-
